1. Adds Counter in LiveView
We test-drive a basic counter in LiveView.

The counter is the first of the seven GUI tasks.

The idea of the exercise is to understand the basic ideas of a
language/toolkit. In our case, we can see how to render a live view,
attach `phx-` event handlers to elements, and how to handle events.

We use a `phx-click` attribute on a button, since there doesn't seem to
be need for more complexity.

7 GUI tasks: https://eugenkiss.github.io/7guis/tasks.