2. Temperature Converter in LiveView
We create a Temperature Converted in LiveView for the 2nd task of the 7
GUIs.

The challenge is to create a UI with bidirectional data flow,
and user-provided text input.

When a user inputs a temperature in Celsius, we transform it into
Fahrenheit and vice versa.

> The formula for converting a temperature C in Celsius into a
temperature F in Fahrenheit is C = (F - 32) * (5/9) and the dual
direction is F = C
* (9/5) + 32.

The task also specifies that non-numerical values should not update the
opposite temperature. So we add some validations to ensure we can parse
the string into an integer, and render an error message otherwise.

All the work is done as soon as the values are changed, so we use
LiveView's `phx-change` attribute. There is no form sumission even
though we use a form element.

7 GUIs: https://eugenkiss.github.io/7guis/tasks