4: Timer in LiveView
What changed?
=============

We create a Timer in LiveView for the fourth task in the [7 GUIs].

The task information:

> The task is to build a frame containing a gauge G for the elapsed time
e, a label which shows the elapsed time as a numerical value, a slider S
by which the duration d of the timer can be adjusted while the timer is
running and a reset button R. Adjusting S must immediately reflect on d
and not only when S is released. It follows that while moving S the
filled amount of G will (usually) change immediately. When e ≥ d is true
then the timer stops (and G will be full). If, thereafter, d is
increased such that d > e will be true then the timer restarts to tick
until e ≥ d is true again. Clicking R will reset e to zero.

> Timer deals with concurrency in the sense that a timer process that
updates the elapsed time runs concurrently to the user’s interactions
with the GUI application. This also means that the solution to competing
user and signal interactions is tested. The fact that slider adjustments
must be reflected immediately moreover tests the responsiveness of the
solution. A good solution will make it clear that the signal is a timer
tick and, as always, has not much scaffolding.

> Challenges: concurrency, competing user/signal interactions,
responsiveness.

[7 GUIs]: https://eugenkiss.github.io/7guis/tasks

Implementation notes
--------------------

I tried to keep this solution simple, and what should no longer be
surprising continues to surprise me -- Elixir & LiveView made this easy.

The task requires a gauge and a slider. For a gauge, I used the `<meter>`
HTML element. For the slider I used an `<input type="range">`.

I built `TimerLive` so it would not really depend on an internal timer.
I did that by creating a `TimerLive.tick/1` function that causes the
timer to move by 1 second. In the end, to keep things simple, I
introduced a timer via `:timer.send_interval/2` in the TimerLive itself.
But I could have made that an external timer (e.g. another GenServer
process).

Most of the implementations is straightforward. The "Reset" button uses
a `phx-click`, and we track the elapsed time (in seconds) through an
assign. The trickier part was making sure the elapsed time was very
responsive. In particular, this requirement from the task was
interesting:

> Adjusting S must immediately reflect on d and not only when S is
released.

So as we adjust the duration slider, we should immediately update the
elapsed time relative to the duration. It turns out that the best way to
do that with an `input` of `type="range"` is to use the `input` event.
From [MDN docs]:

> Note: The input event is fired every time the value of the element
changes. This is unlike the change event, which only fires when the
value is committed, such as by pressing the enter key, selecting a value
from a list of options, and the like.

Unfortunately, LiveView doesn't have a `phx-input` handler, so I
couldn't just rely on that. Thankfully, LiveView has easy
interoperability with JavaScript via hooks.

So, I created a `Slider` hook that adds an event handler for `input`
events and pushes an "update-duration" event with the value.

To keep things simple, and because things work well locally, I did not
add any debouncing, which we could do in JS or with phx-debounce if I
used something like phx-change.

[MDN docs]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/input_event