WIP: Flesh out test cases, add crud_users table
We leave the update test failing.

I'm not yet sure how we can best test this. We can have more than one
submit button for a form if we're doing regular HTTP form submissions.

We can do it via the `formaction` attribute on the submit button. But
LiveViewTest only supports a single `render_submit`. So I'm not sure if
we'll:

- (a) pass the data into `render_submit` instead of using `form/3`, or
- (b) pass the data into a `render_click` which targets the update
  element.

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/form#attr-action