Remove need for filtered_users assign (invalid states)
We remove the need for filtered_users assign. In the first
implementation, we were keeping track of users twice: `users` and
`filtered_users`. We did that so we could filter the list of users and
at the same time keep the canonical list of users if the text filter
changed.

Well, in an effort to think about removing impossible states, I knew
that we didn't want two lists of users which could somehow get out of
sync (and which we had to keep in sync after ever create/update/delete
action). That's when I realized we could keep the search term in assigns
and filter the list when rendered. That way we only need one list of
users.