Refactoring: remove sync selected_user and user_params
What changed?
=============

There was something bothering me about the fact that when we selected a
user we had to update the `selected_user` and the `user_params`. And the
same was true of resetting the selected user (after delete).

It seemed to me like related data that ideally should never be out of
sync. So what to do?

We remove the need to synchronize them by changing `user_params` to hold
a user struct. When we create or update, we grab that user and make it
into a map to pass as params. But by allowing the `current_user` to be a
full user, we can remove the need for a `selected_user`. Instead, we
infer if a user has been selected based on whether or not the user has
an id.