Try out using a sum type for changes & errors
We try using a sum type for a user's changes and errors.

So the assign `:user_changes` could be one of the following:

```elixir
{:new_changes, changes} | {:valid_changes, changes} | {:invalid_changes,
changes, errors}
```

Whenever a valid or invalid change is made, we reset the changes to
`:new_changes`, meaning they haven't been validated. We only validate
when we want, which is before creating the user but not when setting the
first or last names.

This removes the need to keep errors in sync with the user. Now, we
cannot have a UI that has a first & last name but also displays the
"can't be blank" error (which was possible with our previous domain
model).