Update several dependencies
We wanted to use Phoenix LiveView 0.17. While we're updating that
dependency, and since this isn't a production system, we update as many
as we can by unlocking all of our dependencies. We have to bump
`phoenix_html` and `phoenix` versions to resolve conflicts, and we also
resolve any npm vulnerabilities that could be fixed automatically.