6. CircleDrawer in LiveView
We implement the CircleDrawer task in LiveView from the [7 GUI tasks].

> The task is to build a frame containing an undo and redo button as
well as a canvas area underneath. Left-clicking inside an empty area
inside the canvas will create an unfilled circle with a fixed diameter
whose center is the left-clicked point. The circle nearest to the mouse
pointer such that the distance from its center to the pointer is less
than its radius, if it exists, is filled with the color gray. The gray
circle is the selected circle C. Right-clicking C will make a popup menu
appear with one entry “Adjust diameter..”. Clicking on this entry will
open another frame with a slider inside that adjusts the diameter of C.
Changes are applied immediately. Closing this frame will mark the last
diameter as significant for the undo/redo history. Clicking undo will
undo the last significant change (i.e. circle creation or diameter
adjustment). Clicking redo will reapply the last undoed change unless
new changes were made by the user in the meantime.

> Circle Drawer’s goal is, among other things, to test how good the
common challenge of implementing an undo/redo functionality for a GUI
application can be solved. In an ideal solution the undo/redo
functionality comes for free resp. just comes out as a natural
consequence of the language / toolkit / paradigm. Moreover, Circle
Drawer tests how dialog control*, i.e. keeping the relevant context
between several successive GUI interaction steps, is achieved in the
source code. Last but not least, the ease of custom drawing is tested.

[7 GUI tasks]: https://eugenkiss.github.io/7guis/tasks/

Implementation notes
--------------------

There are several ways we could have accomplished this task. The first
step was to draw circles. At first, I considered using a canvas and
rendering 2D circles.

Some helpful resources for that:

- https://developer.mozilla.org/en-US/docs/Web/API/Canvas_API/Tutorial/Drawing_shapes
- https://developer.mozilla.org/en-US/docs/Web/API/CanvasRenderingContext2D/arc

But in the end, I opted for rendering the circles within an `<svg>`
tag with `<circle>` elements.

That allowed me to move the circle knowledge to the server (which
circles exist and which is selected). So, instead of JavaScript being
responsible for drawing the circles, we let the server define them by
iterating over a list of circle we have in our assigns.

Keeping track of circles
------------------------

We keep track of the circles in a `%Canvas{}` struct. Each circle is a
`%Circle{}` struct with `x` and `y` coordinates and radius `r`.

When we send a "canvas-click" event from the front-end, we check if the
coordinates land within an existing circle. We do that with a private
helper function: `CircleDrawer.within_circle?/2`.

If the coordinates land in a place where no circle exists, we add a new
circle to the list. If a circle exists, we mark the circle as selected
and fill it with a different color.

Undo and redo operations
------------------------

One of the main aims of the task is to be able to undo and redo
operations. I originally thought I'd do that by keeping a list of
operations that had been done and then reapplying the transformations
needed. For example, you could image it as a stack of operations:

```
[
  {:add_circle, {x, y, r}}
  {:change_radius, {x, y, r}, new_r}
  ...

  etc

  ...
  {:select_circle, {x, y, r}}
]
```

But then, I realized there was an easier way. Our `canvas` structs are
immutable (thanks Elixir!). So every time we add a circle, remove one,
or change the radius, we create a new canvas. We can use that to our
advantage.

For undo and redo, all we have to do is keep the list of old canvases
that came before. So, for example, you can consider the history of
canvases as a stack:

```
[
  %Canvas{circles: [], selected: nil]},
  %Canvas{circles: [%Circle{x, y, r}], selected: nil]},
  %Canvas{circles: [%Circle{x, y, r}, %Circle{x, y, r}], selected: nil]},
  ...

  etc

  ...
  %Canvas{circles: [%Circle{x, y, r}, ...], selected: %Circle{x, y, r}]},
]

```

Then, when we want to undo something, we grab the last canvas in the
undo history and store the current canvas in the redo history. If we
want to redo an action, we grab the last canvas from the redo history
and store the current canvas in the undo history.

JavaScript interop
-----------------

For this task, I knew we would have to rely on JavaScript
interoperability a bit. There are several ways in which we could have
interacted with JavaScript. This is where I landed:

The SVG canvas has a `phx-hook=CircleDrawer` attached to it. We add an
even listener to that element on `mounted()` so that a "click" on the
canvas is handled by our JavaSript Hook. We need to do this because we
cannot simply rely on `client_x` and `client_y` coordinates (mouse
coordinates) to draw our circles. If we could, we could include that
data in the "click" event metadata. (see:
https://hexdocs.pm/phoenix_live_view/bindings.html#click-events)

Instead, we have to translate the client's x and y coordinates into SVG
coordinates. That's what our phx hook does. There's a matrix transform
function that we can access via JavaScript that can translate our
event's coordinates into SVG coordinates. Then, our JavaScript sends a
"canvas-click" event (via `pushEvent`) to the server with the
coordinates that we care about.

We also use a `SelectedCircle` hook to add an event listener for the
"contextmenu" event. The task specifies that a right click on the
selected circle should pop up a menu. The only way (I know) to react to
a right click is through the "contextmenu" event. So, we listen for that
event and open the modal when it happens. It's worth noting that this is
something we don't want to do with LiveView if possible. Since all we
want to do is open a modal, it's ideal if we can keep it all on the
client side instead of having to go to the server just to open the
modal. That modal has the radius-adjustment UI.

We then have a separate `phx-hook=CircleDiameterSlider` hook. That hook
adds and event listener for "input" so we can adjust the radius of the
`<circle>` element.

One thing I wanted to ensure here was that we'd handle the radius
changes purely in JS at first. You can imagine a user adjusting that to
see what looks like a good size of for the circle. It's only when we
close the modal that we should "save" the new radius for the circle. So,
it's only then that we send an event to the server with the final
radius. We accomplish that by NOT sending events when the "input"
changes. Instead, when the modal is closed, we use `LiveView.JS` to
`dispatch` a client-side event to the same element, and our
`CircleDiameterSlider` hook then sends an event to the server with the
final radius.