Move circle knowledge to server
We move the knowledge of when to draw a circle to the server

Roughly speaking, my guess is that we should keep track of a map of
circles where the key is a map of coordinates and the value is the
radius.

When we send a "canvas-click" event from the front-end, we reference
that map to see if we already have a circle in those coordinates (+ the
radius -- not sure how to do that yet).

If the circle doesn't exist, we send a "draw-circle" event. If it
exists, we send "fill-circle" with the circle's coordinates + radius.

We'll also want to keep track of all the Undo operations (probably in a
list) and Redo operations (probably in another list). We can then push
and pop operations as needed.